Toggle lights

+++ include/globals.hpp
@@ -7,7 +7,7 @@
 
 namespace Global
 {
-    constexpr int WIZ_UDP_BROADCAST_PORT {38899};
+    constexpr u_int16_t WIZ_UDP_BROADCAST_PORT {38899};
 
     inline Wiz::Controller* wizController = nullptr;
     inline Logger logger;
